Singleton pattern added, minor formatting in AbstractFactory.md/Builder.md
