+++ docs/AbstractFactory.md
@@ -1,3 +1,4 @@
+Abstract Factory sources:
 https://www.youtube.com/watch?v=04J_fL5zg3U&t=492s
 https://sbcode.net/python/factory/
 https://refactoring.guru/pl/design-patterns/abstract-factory
