+++ docs/Builder.md
@@ -2,5 +2,3 @@ Builder sources:
 https://www.youtube.com/watch?v=bV1jSlPvtog
 https://medium.com/design-patterns-in-python/builder-design-pattern-in-python-9a66de3e000b
 https://refactoring.guru/pl/design-patterns/builder/python/example
-
-
