+++ docs/Singleton.md
@@ -1,0 +1,3 @@
+Singleton sources:
+https://www.youtube.com/watch?v=Qb4rMvFRLJw
+https://refactoring.guru/pl/design-patterns/singleton
